catalogo con map y promises
